+++ spine-ios/Example/Spine iOS Example/SpineExampleApp.swift
@@ -6,9 +6,16 @@
 //
 
 import SwiftUI
+import Spine
 
 @main
 struct SpineExampleApp: App {
+    
+    init() {
+        let version = Spine.version
+        print("Spine \(version)")
+    }
+    
     var body: some Scene {
         WindowGroup {
             ContentView()