+++ spine-ios/Sources/SpineWrapper/include/spine-cpp-light.h
@@ -1,0 +1,1 @@
+./../../../../spine-cpp/spine-cpp-lite/spine-cpp-light.h