+++ spine-ios/CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Add `spine-iOS` SPM package & example app (#1)
 - Basic Mesh Rendering (#2)
+- Spine C++ Swift Wrapper (#3)